+++ data/Prediction_history.csv
@@ -2,3 +2,5 @@ gender,SeniorCitizen,Partner,Dependents,tenure,PhoneService,MultipleLines,Intern
 Male,0,Yes,Yes,0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-month,Yes,Electronic check,18,19,2024-03-12,Logistic Model,Yes,62.16%
 Male,0,Yes,Yes,0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-month,Yes,Electronic check,18,19,2024-03-12,Logistic Model,Yes,62.16%
 Male,0,Yes,Yes,0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-month,Yes,Electronic check,18,19,2024-03-12,Logistic Model,Yes,62.16%
+Female,0,Yes,Yes,0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-month,Yes,Electronic check,18,19,2024-03-14,Logistic Model,Yes,62.16%
+Female,1,Yes,Yes,0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-month,Yes,Electronic check,18,19,2024-03-14,Logistic Model,Yes,62.16%
